requirements.txt
alembic==1.4.3
aniso8601==8.1.0
antiorm==1.2.1
api==0.0.7
APScheduler==3.7.0
atomicwrites==1.4.0
attrs==20.3.0
Authlib==0.15.3
bunch==1.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
coverage==5.4
crayons==0.4.0
cryptography==3.4.6
DateTime==4.3
db==0.1.1
Flask==1.1.2
Flask-APScheduler==1.12.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-OAuth==0.12
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
httplib2==0.19.0
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
nose==1.3.7
numpy==1.20.1
oauth2==1.9.0.post1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.0
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pyodbc==4.0.30
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
schedule==1.1.0
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
webdriver-manager==3.3.0
Werkzeug==1.0.1
zipp==3.4.0
zope.interface==5.2.0
